+++ src/pint/models/solar_system_shapiro.py
@@ -114,7 +114,7 @@ class SolarSystemShapiro(DelayComponent):
                 grp["obs_sun_pos"], psr_dir, self._ss_mass_sec["sun"]
             )
             if self.PLANET_SHAPIRO.value:
-                for pl in ("jupiter", "saturn", "venus", "uranus"):
+                for pl in ("jupiter", "saturn", "venus", "uranus", "neptune"):
                     delay[loind:hiind] += self.ss_obj_shapiro_delay(
                         grp["obs_" + pl + "_pos"], psr_dir, self._ss_mass_sec[pl]
                     )